add reprocessed files Sn116/117 to fix MCNP issue
The rejection of close gamma lines by NJOY in combination
with a bug in NJOY lead to an incorrect ace file for Sn116
and Sn117. The rejection threshold has been reduced and
NJOY patched to fix the issue. The patched NJOY version
has been used to reprocess Sn116 and Sn117 ENDF files.